+++ frontend/src/scenes/onboarding/frameworks/CodeSnippet.tsx
