+++ frontend/src/scenes/onboarding/frameworks/JavaScriptInstructions.tsx
