+++ frontend/src/scenes/onboarding/frameworks/RubyInstructions.tsx
