+++ frontend/src/scenes/onboarding/panels/AutocapturePanel.tsx
